Fixes: tests, tols, type_as->to, solver init, misc.

+++ examples/odenet_mnist.py
@@ -121,7 +121,7 @@ class ODEBlock(nn.Module):
         self.integration_time = torch.tensor([0, 1]).float()
 
     def forward(self, x):
-        self.integration_time = self.integration_time.type_as(x)
+        self.integration_time = self.integration_time.to(x.dtype)
         out = odeint(self.odefunc, x, self.integration_time, rtol=args.tol, atol=args.tol)
         return out[1]
